tidy up - delete SupabaseProvider, move mutate functions & types
