+++ src/index.tsx
@@ -4,9 +4,6 @@ export { SupabaseUserGlobalContextMeta } from "./components/SupabaseUserGlobalCo
 export { SupabaseProvider } from "./components/SupabaseProvider";
 export { SupabaseProviderMeta } from "./components/SupabaseProvider/registerComponentMeta";
 
-export { SupabaseProviderNew } from "./components/SupabaseProviderNew";
-export { SupabaseProviderNewMeta } from "./components/SupabaseProviderNew/registerComponentMeta";
-
 export { SupabaseUppyUploader } from "./components/SupabaseUppyUploader";
 export { SupabaseUppyUploaderMeta } from "./components/SupabaseUppyUploader/registerComponentMeta";
